Fix issue #23 and a lot of changes
-Connection done before the build_exec, for the case "<<eof"
-Reorganization of some functions and some files in builtins folder
-Return status update is done (issue #23)
-New check, to know if the exec exists in check_builtins and
is_builtins
